% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DaMiR.R
\name{DaMiR.Allplot}
\alias{DaMiR.Allplot}
\title{Quality assessment and visualization of expression data}
\usage{
DaMiR.Allplot(data, df, type = c("spearman", "pearson"))
}
\arguments{
\item{data}{Matrix of normalized expression data, i.e transformed counts by vst or rlog}

\item{df}{A data frame with class and known covariates; at least one column with
'class' label must be included}

\item{type}{A character string specifing the metric to be applied to correlation
analysis. Either "spearman" or "pearson" is allowed; default is "spearman"}
}
\value{
A dendrogram and heatmap, MDS plot(s) and a RLE boxplot
}
\description{
This is a helper function to easily draw (1) clustering dendrogram and
heatmap of a sample-per-sample correlation matrix, (2)
multidimensional scaling plots (MDS)  and (3) relative log expression (RLE) boxplots
of expression data.
}
\details{
Please be sure that NAs are not present in \code{df}'s columns.
Plots will not be drawn in the presence of NAs.
}
\examples{
data(DaMiR.Data.Ex.Norm)
data(DaMiR.Data.Ex.dfClCov)
DaMiR.Allplot(DaMiR.Data.Ex.Norm, DaMiR.Data.Ex.dfClCov)
DaMiR.Allplot(DaMiR.Data.Ex.Norm, DaMiR.Data.Ex.dfClCov, type = "pearson")

}
\author{
Mattia Chiesa, Luca Piacentini
}

