% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DaMiR.R
\name{DaMiR.Clustplot}
\alias{DaMiR.Clustplot}
\title{Expression data clustering and heatmap}
\usage{
DaMiR.Clustplot(data, df, type_row = c("euclidean", "correlation"),
  type_col = c("euclidean", "correlation"))
}
\arguments{
\item{data}{A data frame or matrix of normalized expression data, i.e transformed counts by vst or rlog}

\item{df}{A data frame with class and known covariates; at least one column with
'class' label must be included}

\item{type_row}{The metric to be used to cluster rows. Either "euclidean" or "correlation"
is allowed; default is "euclidean"}

\item{type_col}{The metric to be used to cluster cols. Either "euclidean" or "correlation"
is allowed; default is "euclidean"}
}
\value{
A clustering dendrogram and heatmap.
}
\description{
The function helps to draw a clustering dendrogram and a heatmap of expression data.
}
\examples{
data(DaMiR.Data.Ex.t.Norm)
data(DaMiR.Data.Ex.dfClCov)
DaMiR.Clustplot(DaMiR.Data.Ex.t.Norm, DaMiR.Data.Ex.dfClCov)

}
\author{
Mattia Chiesa, Luca Piacentini
}

