% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{DaMiR.Data.Ex.dfClCov}
\alias{DaMiR.Data.Ex.dfClCov}
\title{Example gene-expression dataset for DaMiRseq package}
\format{A dataframe with 80 samples (rows) and 5 variables(columns):
\describe{
 \item{center}{center where sample has been collected}
 \item{sex}{sample's gender}
 \item{age}{sample's age}
 \item{death}{kind of sample's death, based on Hardy scale}
 \item{class}{sample's class}
}}
\usage{
DaMiR.Data.Ex.dfClCov
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A dataset with class and covariates information
}
\keyword{datasets}

