% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{DaMiR.Data.Ex.Norm}
\alias{DaMiR.Data.Ex.Norm}
\title{Example gene-expression dataset for DaMiRseq package}
\format{A matrix with 13998 genes (rows) and 80 samples (columns):}
\usage{
DaMiR.Data.Ex.Norm
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A dataset with count matrix to test several DaMiRseq functions
}
\keyword{datasets}

