% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{DaMiR.Data.Ex.relief}
\alias{DaMiR.Data.Ex.relief}
\title{Example ranking dataset for DaMiRseq package}
\format{A dataframe with 160 genes (rows) and 2 variables(columns):
\describe{
 \item{reliefF Score}{reliefF score for each gene}
 \item{scaled reliefF Score}{scaled reliefF score for each gene, by z-score}
}}
\usage{
DaMiR.Data.Ex.relief
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A dataset with relieF and scaled reliefF scores for each gene
}
\keyword{datasets}

