% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{DaMiR.Data.Ex.SV}
\alias{DaMiR.Data.Ex.SV}
\title{Example Surrogate Variables dataset for DaMiRseq package}
\format{A matrix with 80 samples (rows) and 12 Surrogate Variables (columns):}
\usage{
DaMiR.Data.Ex.SV
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A dataset with Surrogate Variables to test DaMiRseq functions
}
\keyword{datasets}

