% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{DaMiR.Data.Ex.t.Norm}
\alias{DaMiR.Data.Ex.t.Norm}
\title{Example gene-expression dataset for DaMiRseq package}
\format{A dataframe with 80 samples (rows) and 160 variables(genes):}
\usage{
DaMiR.Data.Ex.t.Norm
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A dataset with normalized values to test feature selection in DaMiRseq package
}
\keyword{datasets}

