% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classif_2_Classes_Predict.R
\name{DaMiR.EnsembleLearning2cl_Predict}
\alias{DaMiR.EnsembleLearning2cl_Predict}
\title{Predict new samples class}
\usage{
DaMiR.EnsembleLearning2cl_Predict(data, models_List)
}
\arguments{
\item{data}{A data frame of normalized expression data.
Rows and Cols should be, respectively, observations and features}

\item{models_List}{A list with the models trained by
\link{DaMiR.EnsembleLearning2cl_Training} function.}
}
\value{
A dataframe containing the predictions
}
\description{
The models learned by the
\link{DaMiR.EnsembleLearning2cl_Training} functionn are applied to
 a dataset, in order to predict the samples class
}
\details{
This function implements the test step of
\link{DaMiR.EnsembleLearning2cl} function
}
\examples{
# use example data:
data(selected_features)
data(df)
set.seed(1)
# only for the example:
# speed up the process setting a low 'iter' argument value;
# for real data set use default 'iter' value (i.e. 100) or higher:
#  Tr_res <- DaMiR.EnsembleLearning2cl_Training(
#  selected_features,classes=df$class, fSample.tr.w=0.6, iter=3,
# cl_type=c("RF","LR"))
# DaMiR.EnsembleLearning2cl_Predict(selected_features, Tr_res)

}
\author{
Mattia Chiesa, Luca Piacentini
}
