% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classif_2_Classes_Test.R
\name{DaMiR.EnsembleLearning2cl_Test}
\alias{DaMiR.EnsembleLearning2cl_Test}
\title{Test Binary Classifiers}
\usage{
DaMiR.EnsembleLearning2cl_Test(data, classes, models_List)
}
\arguments{
\item{data}{A data frame of normalized expression data.
Rows and Cols should be, respectively, observations and features}

\item{classes}{A class vector with \code{nrow(data)} elements.
Each element represents the class label for each observation.
Two different class labels are allowed}

\item{models_List}{A list with the models trained by
\link{DaMiR.EnsembleLearning2cl_Training} function.}
}
\value{
A dataframe containing the predictions on the testset
}
\description{
This function tests the models learned by the
\link{DaMiR.EnsembleLearning2cl_Training} function, on a test set
}
\details{
This function implements the test step of
\link{DaMiR.EnsembleLearning2cl} function
}
\examples{
# use example data:
data(selected_features)
data(df)
set.seed(1)
# only for the example:
# speed up the process setting a low 'iter' argument value;
# for real data set use default 'iter' value (i.e. 100) or higher:
#  Tr_res <- DaMiR.EnsembleLearning2cl_Training(
#  selected_features,classes=df$class, fSample.tr.w=0.6, iter=3,
# cl_type=c("RF","LR"))
# DaMiR.EnsembleLearning2cl_Test(selected_features, classes=df$class,
# Tr_res)

}
\author{
Mattia Chiesa, Luca Piacentini
}
