% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classif_2_Classes_Predict.R
\name{DaMiR.EnsL_Predict}
\alias{DaMiR.EnsL_Predict}
\title{Predict new samples class}
\usage{
DaMiR.EnsL_Predict(data, bestModel)
}
\arguments{
\item{data}{A SummarizedExperiment object or a data frame/matrix
of normalized expression data. Rows and Cols should be
observations and features, respectively.}

\item{bestModel}{The best model, selected between those trained by
the \link{DaMiR.EnsL_Train} function.}
}
\value{
A matrix containing the predictions
}
\description{
The best model learned by the
\link{DaMiR.EnsL_Train} functionn is tested on
 a new dataset, in order to predict the samples class
}
\details{
This function implements the prediction step on new data,
given a model learned by \link{DaMiR.EnsL_Train}
}
\examples{
# use example data:
data(selected_features)
data(df)

}
\author{
Mattia Chiesa, Luca Piacentini
}
