% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DaMiR.goldenDice}
\alias{DaMiR.goldenDice}
\title{Generate a Number to Set Seed}
\usage{
DaMiR.goldenDice()
}
\value{
An integer number.
}
\description{
This function implements a formula based on current
date and time.
}
\details{
The number is generated by combining current seconds (S),
minutes (Mi), hours (H), days (D), months (Mo), years (Y) and golden
ratio (\eqn{\phi}), in the form:

\deqn{ Num = (S * Mi + H * D * Mo / D) ^ \phi}
}
\examples{
gen_numb <- DaMiR.goldenDice()
set.seed(gen_numb)

}
\author{
Mattia Chiesa, Luca Piacentini
}

