% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DaMiR.R
\name{DaMiR.MDSplot}
\alias{DaMiR.MDSplot}
\title{Plot multidimentional scaling (MDS)}
\usage{
DaMiR.MDSplot(data, df, type = c("spearman", "pearson"))
}
\arguments{
\item{data}{A data frame or matrix of normalized expression data, i.e. transformed counts by vst or rlog}

\item{df}{A data frame with class and known covariates; at least one column with
'class' label must be included}

\item{type}{A character string specifing the metric to be applied to correlation
analysis. Either "spearman" or "pearson" is allowed; default is "spearman"}
}
\value{
A MDS plot, using only 'class' information
}
\description{
A MDS plot is drawn in order to visualize class clustering.
}
\details{
The MDS plot is drawn taking as input a dissimilarity matrix produced by either
a sample-per-sample Pearson's or Spearman's correlation of normalized expression data.
}
\examples{
data(DaMiR.Data.Ex.t.Norm)
data(DaMiR.Data.Ex.dfClCov)
DaMiR.MDSplot(DaMiR.Data.Ex.t.Norm, DaMiR.Data.Ex.dfClCov)

}
\author{
Mattia Chiesa, Luca Piacentini
}

