% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_norm_DaMiR.R
\name{DaMiR.normalization}
\alias{DaMiR.normalization}
\title{Filter non Expressed and 'Hypervariant' features and Data Normalization}
\usage{
DaMiR.normalization(dds, minCounts = 10, fSample = 0.5, hyper = c("yes",
  "no"), th.cv = 3, type = c("vst", "rlog"))
}
\arguments{
\item{dds}{A \code{DESeq} object}

\item{minCounts}{Minimum reads counts; default is 10}

\item{fSample}{Fraction of samples with \code{minCounts} counts; default is 0.5}

\item{hyper}{Flag to enable gene filtering by Coefficient of
Variation (CV); default is "yes"}

\item{th.cv}{Threshold of minimum CV to consider a feature 'Hypervariant'
accross samples; default is 3}

\item{type}{Type of normalization to be applied: \code{varianceStabilizingTransformation}
(\code{vst}) or \code{rlog} are allowed; default is "\code{vst}"}
}
\value{
A normalized expression matrix.
}
\description{
Features will be firstly filtered based on their expression value and/or by
 their variability across samples; features will be then normalized.
}
\details{
Before normalization step, this function allows the user to filter features by:
\itemize{
  \item Expression - Features will be filtered out whether their reads count
  do not reach a \code{minCounts} in at least \code{fSample} of samples;
  \item CV - The CV of each feature is individually calculated for each sample class.
  Featurers with both class CV greater than \code{th.cv} will be discarded.
  Computing a class restricted CV may prevent the removal of hypervariant features that
  may be specifically associated with a certain class. This could be important, for example, for
  immune genes whose expression under definite conditions may unveil peculiar class-gene
  association.
}
Finally, expressed features will be normalized by \code{varianceStabilizingTransformation}
 (default) or \code{rlog}, both implemented in \code{DESeq2} package. We suggest to
 use \code{varianceStabilizingTransformation} to speed up the normalization process
 because \code{rlog} is very time-consuming despite the two methods produce quite
 similar results.
}
\examples{
rawdata.path <- system.file(package = "DaMiRseq","extdata")
filecounts <- list.files(rawdata.path,full.names = TRUE)[1]
filecovariates <- list.files(rawdata.path,full.names = TRUE)[2]
list.res <- DaMiR.read(filecounts,filecovariates)
dds <- list.res$dds

data_norm <- DaMiR.normalization(dds)
data_norm <- DaMiR.normalization(dds, fSample = 0.6)
data_norm <- DaMiR.normalization(dds, th.cv = 2.5, type = "rlog")

}
\author{
Mattia Chiesa, Luca Piacentini
}
\references{
Michael I Love, Wolfgang Huber and Simon Anders (2014): Moderated estimation of
 fold change and dispersion for RNA-Seq data with DESeq2. Genome Biology
}
\seealso{
\code{\link{varianceStabilizingTransformation}, \link{rlog} }
}

