% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcount_DaMiR.R
\name{DaMiR.read}
\alias{DaMiR.read}
\title{Import RNA-Seq count data and covariates}
\usage{
DaMiR.read(counts_file, classcov_file)
}
\arguments{
\item{counts_file}{The name (with path, if needed) of a tab-delimited file which contains
RNA-Seq count data. Each row is a feature (i.e. gene, transcript, exon etc.) and
each column is a sample}

\item{classcov_file}{The name (with path, if needed) of a tab-delimited file which
contains experiment information. Each row is a sample and each column is a variable.
This file must contain at least one column which represent 'class' information for data
adjustment and classification; the class column must be labeled as 'class'}
}
\value{
A list containing:
\itemize{
  \item A \code{DESeqDataSet} object as created by DESeqDataSetFromMatrix function of DESeq2.
  \item A data frame with class and covariates information.
}
}
\description{
This function allows the user to simultaneously import counts, class (mandatory) and
covariate (optional) data, and creates a \code{DESeqDataSet} object.
}
\examples{
rawdata.path <- system.file(package = "DaMiRseq","extdata")
filecounts <- list.files(rawdata.path,full.names = TRUE)[1]
filecovariates <- list.files(rawdata.path,full.names = TRUE)[2]
list.res <- DaMiR.read(filecounts,filecovariates)

}
\author{
Mattia Chiesa, Luca Piacentini
}
\references{
Michael I Love, Wolfgang Huber and Simon Anders (2014): Moderated estimation of
 fold change and dispersion for RNA-Seq data with DESeq2. Genome Biology
}
\seealso{
\code{\link{DESeqDataSetFromMatrix}}
}

