% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_norm_DaMiR.R
\name{DaMiR.sampleFilt}
\alias{DaMiR.sampleFilt}
\title{Filter Samples by Mean Correlation Distance Metric}
\usage{
DaMiR.sampleFilt(data, df, th.corr = 0.9, type = c("spearman", "pearson"))
}
\arguments{
\item{data}{Matrix of normalized expression data, i.e transformed counts by vst or rlog.
A log2 transformed expression matrix is also accepted}

\item{df}{A data frame with known covariates; at least one column with 'class'
label must be included}

\item{th.corr}{Threshold of mean correlation; default is 0.9}

\item{type}{Type of correlation metric; default is "spearman"}
}
\value{
A list containing:
\itemize{
  \item A normalized and filtered expression matrix.
  \item A data frame with class and optional covariates information.
}
}
\description{
This function implements a sample-per-sample correlation. Samples with a mean
correlation lower than a user's defined threshold will be filtered out.
}
\details{
This step introduces a sample quality checkpoint. Global gene expression should,
in fact, exhibit a high correlation among biological replicates; conversely, low correlated
samples may be suspected to bear some technical artifact (e.g. poor RNA or library
preparation quality), despite they may have passed sequencing quality checks. If not assessed,
these samples may, thus, negatively affect all the downstream analysis. This function looks at
the mean absolute correlation of each sample and removes those samples with a mean correlation
lower than the value set in \code{th.corr} argument. This threshold may be specific for
different experimental setting but should be as high as possible. For sequencing data we
suggest to set \code{th.corr} greater than 0.85.
}
\examples{
data(DaMiR.Data.Ex.Norm)
data(DaMiR.Data.Ex.dfClCov)
list.res <- DaMiR.sampleFilt(DaMiR.Data.Ex.Norm, DaMiR.Data.Ex.dfClCov)
list.res <- DaMiR.sampleFilt(DaMiR.Data.Ex.Norm, DaMiR.Data.Ex.dfClCov, th.corr = 0.8)
list.res <- DaMiR.sampleFilt(DaMiR.Data.Ex.Norm, DaMiR.Data.Ex.dfClCov, type = "pearson")

}
\author{
Mattia Chiesa, Luca Piacentini
}

