% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_Batch_DaMiR.R
\name{DaMiR.SVadjust}
\alias{DaMiR.SVadjust}
\title{Remove variable effects from expression data}
\usage{
DaMiR.SVadjust(data, df, sv, n.sv)
}
\arguments{
\item{data}{Matrix of normalized expression data, i.e. transformed counts by vst or rlog.
A log2 transformed expression matrix is also accepted}

\item{df}{A data frame with class and known covariates; at least one column with
'class' label must be included}

\item{sv}{The matrix of surrogate variables identified by \code{\link{DaMiR.SV}} function}

\item{n.sv}{The number of surrogate variables to be used to adjust the data}
}
\value{
A list containing:
\itemize{
  \item A normalized and adjusted expression matrix.
  \item A data frame with class and optional covariates information.
}
}
\description{
This function removes surrogate or other confounding variable effects from
normalized expression data by the usage of \code{\link{removeBatchEffect}} function
 of \code{limma} package.
}
\examples{
data(DaMiR.Data.Ex.Norm)
data(DaMiR.Data.Ex.dfClCov)
data(DaMiR.Data.Ex.SV)
data.adjusted <- DaMiR.SVadjust(DaMiR.Data.Ex.Norm, DaMiR.Data.Ex.dfClCov, sv = DaMiR.Data.Ex.SV)
data.adjusted <- DaMiR.SVadjust(DaMiR.Data.Ex.Norm, DaMiR.Data.Ex.dfClCov, sv = DaMiR.Data.Ex.SV, n.sv = 3)

}
\author{
Mattia Chiesa, Luca Piacentini
}
\seealso{
\code{\link{removeBatchEffect}},
\code{\link{DaMiR.SV}}
}

