% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DaMiR.transpose}
\alias{DaMiR.transpose}
\title{Matrix transposition and replacement of '.' and '-' special
 characters}
\usage{
DaMiR.transpose(data)
}
\arguments{
\item{data}{Matrix of normalized expression data, i.e. transformed
counts by vst or rlog.
A log2 transformed expression matrix is also accepted}
}
\value{
Normalized matrix in which each row is a sample and each
column is a feature
}
\description{
This function transposes matrix and replaces '.' and '-'
 special characters.
}
\examples{
data(data_norm)
data.transposed <- DaMiR.transpose(assay(data_norm))

}
\author{
Mattia Chiesa, Luca Piacentini
}
