% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{data_min}
\alias{data_min}
\title{Example gene-expression dataset for DaMiRseq package}
\format{A data frame with 40 samples (rows) and 87 genes (columns)}
\usage{
data_min
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A dataset with a small dimension of normalized expression data
in DaMiRseq package
}
\keyword{datasets}
