% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{data_norm}
\alias{data_norm}
\title{A dataset with a normalized matrix to test several DaMiRseq functions:
sample data are a subset of Genotype-Tissue Expression (GTEx)
RNA-Seq database (dbGap Study Accession: phs000424.v6.p1)}
\format{A SummarizedExperiment object containing an assay of
4897 genes (rows) and 40 samples (columns) and a colData with 5 variables}
\usage{
data_norm
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A dataset with a normalized matrix to test several DaMiRseq functions:
sample data are a subset of Genotype-Tissue Expression (GTEx)
RNA-Seq database (dbGap Study Accession: phs000424.v6.p1)
}
\keyword{datasets}

