% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{data_reduced}
\alias{data_reduced}
\title{Example gene-expression dataset for DaMiRseq package}
\format{A list with:
\describe{
 \item{data}{reduced expression matrix}
 \item{variables}{a data frame with variables}
}}
\usage{
data_reduced
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A dataset with a small dimension of normalized expression data
in DaMiRseq package
}
\keyword{datasets}
