% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{data_relief}
\alias{data_relief}
\title{Example ranking dataset for DaMiRseq package}
\format{A dataframe with 87 genes (rows) and 2 variables (columns):
\describe{
 \item{reliefF Score}{reliefF score for each gene}
 \item{scaled reliefF Score}{scaled reliefF score for each gene, by z-score}
}}
\usage{
data_relief
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A data frame with relieF and scaled reliefF scores for each gene
}
\keyword{datasets}
