% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{df}
\alias{df}
\title{Example gene-expression dataset for DaMiRseq package}
\format{A dataframe with 40 samples (rows) and 5 variables (columns):
\describe{
 \item{center}{center where sample has been collected}
 \item{sex}{sample's gender}
 \item{age}{sample's age}
 \item{death}{kind of sample's death, based on Hardy scale}
 \item{class}{sample's class}
}}
\usage{
df
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A data frame with class and covariates information
}
\keyword{datasets}
