% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{selected_features}
\alias{selected_features}
\title{Example gene-expression dataset for DaMiRseq package}
\format{A dataframe with 40 samples (rows) and 7 variables (genes):}
\usage{
selected_features
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A dataset with normalized expression data to build classification models
in DaMiRseq package
}
\keyword{datasets}

