% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{SEtest_norm}
\alias{SEtest_norm}
\title{A sample dataset with a normalized count matrix for "testthat"
functions.}
\format{A SummarizedExperiment object containing an assay of
100 genes (rows) and 11 samples (columns) and a colData with 5 variables}
\usage{
SEtest_norm
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A sample dataset with a normalized count matrix for "testthat"
functions.
}
\keyword{datasets}
