% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{sv}
\alias{sv}
\title{Example Surrogate Variables dataset for DaMiRseq package}
\format{A matrix with 40 samples (rows) and 4 surrogate variables (columns):}
\usage{
sv
}
\value{
An example dataset for \code{DaMiRseq} package
}
\description{
A dataset with surrogate variables to test DaMiRseq functions
}
\keyword{datasets}

