% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{DaMiR.Allplot}
\alias{DaMiR.Allplot}
\title{Quality assessment and visualization of expression data}
\usage{
DaMiR.Allplot(data, df, type = c("spearman", "pearson"))
}
\arguments{
\item{data}{A SummarizedExperiment object or an expression matrix}

\item{df}{A data frame with class and known covariates (or a subset of them); at least one column with
'class' label must be included}

\item{type}{A character string specifing the metric to be applied to correlation
analysis. Either "spearman" or "pearson" is allowed; default is "spearman"}
}
\value{
A dendrogram and heatmap, MDS plot(s) and a RLE boxplot
}
\description{
This is a helper function to easily draw (1) clustering dendrogram and
heatmap of a sample-per-sample correlation matrix, (2)
multidimensional scaling plots (MDS)  and (3) relative log expression (RLE) boxplots
of expression data.
}
\details{
Please be sure that NAs are not present in \code{df}'s columns.
Plots will not be drawn in the presence of NAs.
}
\examples{
# use example data:
data(data_norm)
data(df)
# Draw clustering dendrogram and heatmap, MDS, RLE boxplot:
DaMiR.Allplot(data=data_norm, df=df)

}
\author{
Mattia Chiesa, Luca Piacentini
}

